gameworld, otb/map, compositor: Ambient light is now supplied by the MapDataSource.
This is not necessarily the ideal place to provide it; it's more of an aspect of the game world itself -- not the map -- and it's unclear how it would work if the map is assembled for multiple map sources. However, it's a convenient way to pass it into the compositor at this time; it can be moved to a separate parameter at a later time.

In addition, to clean up the compositor package, the compositeTile code has been moved to a separate file -- hopefully enabling some future clarity on how to provide shared layouting code for both image-based and for DOM-based compositor.